co_project/co_proj/requirements.txt
asgiref==3.4.1
captcha==0.3
click==8.0.1
colorama==0.4.4
Django==3.2.7
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-taggit==1.2.0
fields==5.0.0
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-client==0.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
Pillow==8.3.2
pyreadline==2.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1
